factor data handling into container
