+++ src/viz-population-over-time/index.js
@@ -1,1 +1,1 @@
-export { default } from "./VizPopulationOverTime";
+export { default } from "./VizPopulationOverTimeContainer";
